Updated getAll()

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -5,9 +5,12 @@ import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -28,17 +31,24 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public void delete(int id) {
-        repository.remove(id);
+        if (repository.containsKey(id))
+            repository.remove(id);
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        if (repository.containsKey(id))
+            return repository.get(id);
+        return null;
     }
 
+    //будут ли последние по времени записи наверху?
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        if (!repository.isEmpty())
+            return repository.values().stream()
+                    .sorted((Comparator.comparing(Meal::getDateTime)))
+                    .collect(Collectors.toList());
+        return Collections.emptyList();
     }
 }
-
